Various cleanups.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/tiptop/trunk@270 fd2ea14d-3e48-4a98-bac5-13ccbbe52f7f

+++ src/xml-parser.c
@@ -59,10 +59,10 @@ static int is_blank(char* t)
 static void option_update(xmlChar* name, xmlChar* val, struct option* opt)
 {
   if(!xmlStrcmp(name, (const xmlChar *) "delay")) {
-    opt->delay = atof((const char*)val);
+    opt->delay = (float)atof((const char*)val);
   }
   if(!xmlStrcmp(name, (const xmlChar *) "cpu_threshold")) {
-    opt->cpu_threshold = atof((char*)val);
+    opt->cpu_threshold = (float)atof((char*)val);
   }
   if(!xmlStrcmp(name, (xmlChar *) "batch")) {
     opt->batch = (opt->batch || atoi((char*)val));
